Reorganize files to leave all inside meetup module
