+++ src/getUpcomingMeetups.js
@@ -1,17 +1,17 @@
-const meetupAPI = require('./meetupAPI');
-const meetupRenderer = require('./meetupRenderer');
+const api = require('./meetup/api');
+const renderer = require('./meetup/renderer');
 
 function onlyValid(x) {
   return x !== undefined;
 }
 
 function render(meetup) {
-  return meetupRenderer.render(meetup);
+  return renderer.render(meetup);
 }
 
 const getUpcomingMeetups = (meetups, token) => {
   const promises = meetups
-    .map(meetupName => meetupAPI.upcomingMeetup(meetupName, token));
+    .map(meetupName => api.upcomingMeetup(meetupName, token));
 
   return Promise.all(promises)
     .then(response => response.filter(onlyValid))
